fcn-dev - added README info to update the .env files before starting

+++ install.sh
@@ -6,8 +6,12 @@ read -p "Do you meet these prerequisites (Y/n)? " -n 1 -r
 echo    # (optional) move to a new line
 if [[ $REPLY =~ ^[Yy]$ ]]; then
   # do dangerous stuff
-  cp .env.dist client/.env
-  cp .env.dist .env
+  if [ ! -f client/.env ]; then
+    cp .env.dist client/.env
+  fi
+  if [ ! -f .env ]; then
+    cp .env.dist .env
+  fi
   cd client
   sudo npm install react-scripts@2.1.3
   sudo npm install @material-ui/core
